third_party/blink/web_tests/wpt_internal/infrastructure/wpt-server-http.sub.html

<!doctype html>
<title>Basic server-side substituion and testharness</title>
<meta charset="utf-8" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <script>
    setup({ single_test: true });
    assert_true('{{domains[www]}}'.startsWith('www.'));
    assert_true({{ports[http][0]}} > 0);
    done();
  </script>
</body>
